extracted user proposa validation to custom hook
